Ignore events that started before PrivCount was enabled
In an earlier commit, we made it so bytes and cells are only counted if
EnablePrivCount is set. But we were still reporting events that started
before PrivCount was enabled, even when we weren't counting their bytes or
cells.

We fix this by storing a timestamp when EnablePrivCount is set, and ignoring
events created before this timestamp.

This fixes the following issues:
* events were biased towards long-running streams, circuits, and
  connections, because they were more likely to overlap the start of the
  collection
* events that overlapped the start of the collection would have lower byte and
  cell counts than expected, because some of their bytes or cells may have
  been ignored.

Closes #305.