filter out zero packet streams from stream traffic model
Streams that don't tranfer any packets are already filtered out
of the packet model, and circuits without any streams are filtered
out of the stream model.

After this commit, we now filter out streams with no packets from
the stream model so that we are consistent, since these same
streams are also filtered out of the packet model.